requirements.txt
numpy==1.26.4
openai==1.12.0
unstructured==0.12.4
unstructured-client==0.18.0
python-dotenv==1.0.1
openpyxl==3.1.2
unstructured[xlsx]
